test: split supervisor module coverage
