+++ src/supervisor.test.ts
@@ -7,12 +7,8 @@ import path from "node:path";
 import {
   Supervisor,
 } from "./supervisor";
-import { buildChecksFailureContext, buildConflictFailureContext } from "./pull-request-failure-context";
-import { inferStateFromPullRequest } from "./pull-request-state";
-import { localReviewHighSeverityNeedsRetry } from "./review-handling";
 import { recoverUnexpectedCodexTurnFailure } from "./supervisor-failure-helpers";
-import { formatDetailedStatus, summarizeChecks } from "./supervisor-status-rendering";
-import { buildDetailedStatusModel, buildDetailedStatusSummaryLines } from "./supervisor-status-model";
+import { formatDetailedStatus } from "./supervisor-status-rendering";
 import { shouldAutoRetryHandoffMissing } from "./supervisor-execution-policy";
 import { handleAuthFailure } from "./supervisor-failure-helpers";
 import {
@@ -23,7 +19,6 @@ import {
   reconcileStaleActiveIssueReservation,
   reconcileTrackedMergedButOpenIssues,
 } from "./recovery-reconciliation";
-import { configuredBotReviewThreads, manualReviewThreads } from "./supervisor-reporting";
 import { StateStore } from "./state-store";
 import { GitHubIssue, GitHubPullRequest, IssueRunRecord, PullRequestCheck, ReviewThread, SupervisorConfig, SupervisorStateFile } from "./types";
 
@@ -4108,2359 +4103,6 @@ function branchName(config: SupervisorConfig, issueNumber: number): string {
   return `${config.branchPrefix}${issueNumber}`;
 }
 
-test("summarizeChecks treats cancelled runs as waiting, not failing", () => {
-  const checks: PullRequestCheck[] = [
-    { name: "build (ubuntu-latest)", state: "CANCELLED", bucket: "cancel", workflow: "CI" },
-  ];
-
-  assert.deepEqual(summarizeChecks(checks), {
-    allPassing: false,
-    hasPending: true,
-    hasFailing: false,
-  });
-});
-
-test("buildChecksFailureContext ignores cancelled runs", () => {
-  const pr: GitHubPullRequest = {
-    number: 42,
-    title: "Test PR",
-    url: "https://example.test/pr/42",
-    state: "OPEN",
-    createdAt: "2026-03-11T14:00:00Z",
-    isDraft: false,
-    reviewDecision: null,
-    mergeStateStatus: "UNSTABLE",
-    mergeable: "MERGEABLE",
-    headRefName: "codex/issue-42",
-    headRefOid: "deadbeef",
-    mergedAt: null,
-  };
-
-  const checks: PullRequestCheck[] = [
-    { name: "build (ubuntu-latest)", state: "CANCELLED", bucket: "cancel", workflow: "CI" },
-    { name: "build (macos-latest)", state: "SUCCESS", bucket: "pass", workflow: "CI" },
-  ];
-
-  assert.equal(buildChecksFailureContext(pr, checks), null);
-});
-
-test("buildChecksFailureContext preserves failing-check reporting fields", () => {
-  const pr: GitHubPullRequest = {
-    number: 42,
-    title: "Test PR",
-    url: "https://example.test/pr/42",
-    state: "OPEN",
-    createdAt: "2026-03-11T14:00:00Z",
-    isDraft: false,
-    reviewDecision: null,
-    mergeStateStatus: "UNSTABLE",
-    mergeable: "MERGEABLE",
-    headRefName: "codex/issue-42",
-    headRefOid: "deadbeef",
-    mergedAt: null,
-  };
-
-  const checks: PullRequestCheck[] = [
-    {
-      name: "build (ubuntu-latest)",
-      state: "FAILURE",
-      bucket: "fail",
-      workflow: "CI",
-      link: "https://example.test/checks/ubuntu",
-    },
-    {
-      name: "test (macos-latest)",
-      state: "TIMED_OUT",
-      bucket: "fail",
-      workflow: "CI",
-    },
-  ];
-
-  const context = buildChecksFailureContext(pr, checks);
-  assert.equal(context?.category, "checks");
-  assert.equal(context?.summary, "PR #42 has failing checks.");
-  assert.equal(context?.signature, "build (ubuntu-latest):fail|test (macos-latest):fail");
-  assert.equal(context?.command, "gh pr checks");
-  assert.deepEqual(context?.details, [
-    "build (ubuntu-latest) (fail/FAILURE) https://example.test/checks/ubuntu",
-    "test (macos-latest) (fail/TIMED_OUT)",
-  ]);
-  assert.equal(context?.url, "https://example.test/pr/42");
-});
-
-test("buildConflictFailureContext preserves merge-conflict reporting fields", () => {
-  const pr: GitHubPullRequest = {
-    number: 42,
-    title: "Test PR",
-    url: "https://example.test/pr/42",
-    state: "OPEN",
-    createdAt: "2026-03-11T14:00:00Z",
-    isDraft: false,
-    reviewDecision: null,
-    mergeStateStatus: "DIRTY",
-    mergeable: "CONFLICTING",
-    headRefName: "codex/issue-42",
-    headRefOid: "deadbeef",
-    mergedAt: null,
-  };
-
-  const context = buildConflictFailureContext(pr);
-  assert.equal(context.category, "conflict");
-  assert.equal(context.summary, "PR #42 has merge conflicts and needs a base-branch integration pass.");
-  assert.equal(context.signature, "dirty:deadbeef");
-  assert.equal(context.command, "git fetch origin && git merge origin/<default-branch>");
-  assert.deepEqual(context.details, ["mergeStateStatus=DIRTY"]);
-  assert.equal(context.url, "https://example.test/pr/42");
-});
-
 test("supervisor module continues to export the Supervisor class", () => {
   assert.equal(typeof Supervisor, "function");
-});
-
-test("inferStateFromPullRequest routes actionable high local-review retry into local_review_fix", () => {
-  const config = createConfig({
-    localReviewEnabled: true,
-    localReviewPolicy: "block_ready",
-    localReviewHighSeverityAction: "retry",
-  });
-  const record = createRecord({
-    state: "draft_pr",
-    pr_number: 44,
-    local_review_head_sha: "head123",
-    local_review_max_severity: "high",
-    local_review_verified_max_severity: "high",
-    local_review_verified_findings_count: 1,
-    local_review_findings_count: 3,
-    local_review_recommendation: "changes_requested",
-    repeated_local_review_signature_count: 1,
-  });
-  const pr: GitHubPullRequest = {
-    number: 44,
-    title: "Test PR",
-    url: "https://example.test/pr/44",
-    state: "OPEN",
-    createdAt: "2026-03-11T00:00:00Z",
-    isDraft: true,
-    reviewDecision: null,
-    mergeStateStatus: "CLEAN",
-    headRefName: "codex/issue-38",
-    headRefOid: "head123",
-  };
-
-  assert.equal(inferStateFromPullRequest(config, record, pr, [], []), "local_review_fix");
-});
-
-test("inferStateFromPullRequest does not wait for Copilot when no lifecycle signal exists", () => {
-  const config = createConfig({ copilotReviewWaitMinutes: 10 });
-  const now = new Date().toISOString();
-  const record = createRecord({
-    state: "pr_open",
-    review_wait_started_at: now,
-    review_wait_head_sha: "head123",
-  });
-  const pr: GitHubPullRequest = {
-    number: 44,
-    title: "Test PR",
-    url: "https://example.test/pr/44",
-    state: "OPEN",
-    createdAt: now,
-    isDraft: false,
-    reviewDecision: null,
-    mergeStateStatus: "CLEAN",
-    mergeable: "MERGEABLE",
-    headRefName: "codex/issue-38",
-    headRefOid: "head123",
-    mergedAt: null,
-  };
-  const checks: PullRequestCheck[] = [{ name: "build", state: "SUCCESS", bucket: "pass", workflow: "CI" }];
-
-  assert.equal(inferStateFromPullRequest(config, record, pr, checks, []), "ready_to_merge");
-});
-
-test("inferStateFromPullRequest waits briefly after ready-for-review for Copilot request propagation", () => {
-  withStubbedDateNow("2026-03-13T05:42:40Z", () => {
-    const config = createConfig({
-      copilotReviewWaitMinutes: 10,
-      reviewBotLogins: ["copilot-pull-request-reviewer"],
-    });
-    const record = createRecord({
-      state: "pr_open",
-      review_wait_started_at: "2026-03-13T05:42:36Z",
-      review_wait_head_sha: "head123",
-    });
-    const pr: GitHubPullRequest = {
-      number: 44,
-      title: "Test PR",
-      url: "https://example.test/pr/44",
-      state: "OPEN",
-      createdAt: "2026-03-13T05:40:00Z",
-      isDraft: false,
-      reviewDecision: null,
-      mergeStateStatus: "CLEAN",
-      mergeable: "MERGEABLE",
-      headRefName: "codex/issue-38",
-      headRefOid: "head123",
-      mergedAt: null,
-      copilotReviewState: "not_requested",
-      copilotReviewRequestedAt: null,
-      copilotReviewArrivedAt: null,
-    };
-    const checks: PullRequestCheck[] = [{ name: "build", state: "SUCCESS", bucket: "pass", workflow: "CI" }];
-
-    assert.equal(inferStateFromPullRequest(config, record, pr, checks, []), "waiting_ci");
-  });
-});
-
-test("inferStateFromPullRequest does not time out immediately when configured review waiting is disabled", () => {
-  const config = createConfig({
-    copilotReviewWaitMinutes: 0,
-    copilotReviewTimeoutAction: "block",
-    reviewBotLogins: ["copilot-pull-request-reviewer"],
-  });
-  const record = createRecord({
-    state: "waiting_ci",
-    review_wait_started_at: "2026-03-11T00:00:00Z",
-    review_wait_head_sha: "head123",
-  });
-  const pr: GitHubPullRequest = {
-    number: 44,
-    title: "Test PR",
-    url: "https://example.test/pr/44",
-    state: "OPEN",
-    createdAt: "2026-03-11T00:00:00Z",
-    isDraft: false,
-    reviewDecision: null,
-    mergeStateStatus: "CLEAN",
-    mergeable: "MERGEABLE",
-    headRefName: "codex/issue-38",
-    headRefOid: "head123",
-    mergedAt: null,
-    copilotReviewState: "requested",
-    copilotReviewRequestedAt: "2026-03-11T00:05:00Z",
-    copilotReviewArrivedAt: null,
-  };
-  const checks: PullRequestCheck[] = [{ name: "build", state: "SUCCESS", bucket: "pass", workflow: "CI" }];
-
-  assert.equal(inferStateFromPullRequest(config, record, pr, checks, []), "waiting_ci");
-});
-
-test("inferStateFromPullRequest waits briefly after ready-for-review for configured bot request propagation", () => {
-  withStubbedDateNow("2026-03-13T05:42:40Z", () => {
-    const config = createConfig({
-      copilotReviewWaitMinutes: 10,
-      reviewBotLogins: ["chatgpt-codex-connector"],
-    });
-    const record = createRecord({
-      state: "pr_open",
-      review_wait_started_at: "2026-03-13T05:42:36Z",
-      review_wait_head_sha: "head123",
-    });
-    const pr: GitHubPullRequest = {
-      number: 44,
-      title: "Test PR",
-      url: "https://example.test/pr/44",
-      state: "OPEN",
-      createdAt: "2026-03-13T05:40:00Z",
-      isDraft: false,
-      reviewDecision: null,
-      mergeStateStatus: "CLEAN",
-      mergeable: "MERGEABLE",
-      headRefName: "codex/issue-38",
-      headRefOid: "head123",
-      mergedAt: null,
-      copilotReviewState: "not_requested",
-      copilotReviewRequestedAt: null,
-      copilotReviewArrivedAt: null,
-    };
-    const checks: PullRequestCheck[] = [{ name: "build", state: "SUCCESS", bucket: "pass", workflow: "CI" }];
-
-    assert.equal(inferStateFromPullRequest(config, record, pr, checks, []), "waiting_ci");
-  });
-});
-
-test("inferStateFromPullRequest allows merge after the Copilot propagation grace window expires", () => {
-  withStubbedDateNow("2026-03-13T05:42:42Z", () => {
-    const config = createConfig({
-      copilotReviewWaitMinutes: 10,
-      reviewBotLogins: ["copilot-pull-request-reviewer"],
-    });
-    const record = createRecord({
-      state: "pr_open",
-      review_wait_started_at: "2026-03-13T05:42:36Z",
-      review_wait_head_sha: "head123",
-    });
-    const pr: GitHubPullRequest = {
-      number: 44,
-      title: "Test PR",
-      url: "https://example.test/pr/44",
-      state: "OPEN",
-      createdAt: "2026-03-13T05:40:00Z",
-      isDraft: false,
-      reviewDecision: null,
-      mergeStateStatus: "CLEAN",
-      mergeable: "MERGEABLE",
-      headRefName: "codex/issue-38",
-      headRefOid: "head123",
-      mergedAt: null,
-      copilotReviewState: "not_requested",
-      copilotReviewRequestedAt: null,
-      copilotReviewArrivedAt: null,
-    };
-    const checks: PullRequestCheck[] = [{ name: "build", state: "SUCCESS", bucket: "pass", workflow: "CI" }];
-
-    assert.equal(inferStateFromPullRequest(config, record, pr, checks, []), "ready_to_merge");
-  });
-});
-
-test("inferStateFromPullRequest keeps waiting when a configured bot request was observed on the current head but has not arrived", () => {
-  withStubbedDateNow("2026-03-11T00:10:00Z", () => {
-    const config = createConfig({
-      copilotReviewWaitMinutes: 10,
-      reviewBotLogins: ["chatgpt-codex-connector"],
-    });
-    const record = createRecord({
-      state: "waiting_ci",
-      review_wait_started_at: "2026-03-11T00:00:00Z",
-      review_wait_head_sha: "head123",
-      copilot_review_requested_observed_at: "2026-03-11T00:05:00Z",
-      copilot_review_requested_head_sha: "head123",
-    });
-    const pr: GitHubPullRequest = {
-      number: 44,
-      title: "Test PR",
-      url: "https://example.test/pr/44",
-      state: "OPEN",
-      createdAt: "2026-03-11T00:00:00Z",
-      isDraft: false,
-      reviewDecision: null,
-      mergeStateStatus: "CLEAN",
-      mergeable: "MERGEABLE",
-      headRefName: "codex/issue-38",
-      headRefOid: "head123",
-      mergedAt: null,
-      copilotReviewState: "not_requested",
-      copilotReviewRequestedAt: null,
-      copilotReviewArrivedAt: null,
-    };
-    const checks: PullRequestCheck[] = [{ name: "build", state: "SUCCESS", bucket: "pass", workflow: "CI" }];
-
-    assert.equal(inferStateFromPullRequest(config, record, pr, checks, []), "waiting_ci");
-  });
-});
-
-test("inferStateFromPullRequest waits when mixed configured bots include Copilot lifecycle state", () => {
-  withStubbedDateNow("2026-03-11T00:10:00Z", () => {
-    const config = createConfig({
-      copilotReviewWaitMinutes: 10,
-      reviewBotLogins: ["chatgpt-codex-connector", "copilot-pull-request-reviewer"],
-    });
-    const requestedAt = "2026-03-11T00:05:00Z";
-    const record = createRecord({
-      state: "waiting_ci",
-      review_wait_started_at: requestedAt,
-      review_wait_head_sha: "head123",
-    });
-    const pr: GitHubPullRequest = {
-      number: 44,
-      title: "Test PR",
-      url: "https://example.test/pr/44",
-      state: "OPEN",
-      createdAt: "2026-03-11T00:00:00Z",
-      isDraft: false,
-      reviewDecision: null,
-      mergeStateStatus: "CLEAN",
-      mergeable: "MERGEABLE",
-      headRefName: "codex/issue-38",
-      headRefOid: "head123",
-      mergedAt: null,
-      copilotReviewState: "requested",
-      copilotReviewRequestedAt: requestedAt,
-      copilotReviewArrivedAt: null,
-    };
-    const checks: PullRequestCheck[] = [{ name: "build", state: "SUCCESS", bucket: "pass", workflow: "CI" }];
-
-    assert.equal(inferStateFromPullRequest(config, record, pr, checks, []), "waiting_ci");
-  });
-});
-
-test("inferStateFromPullRequest keeps waiting when Copilot review was explicitly requested", () => {
-  withStubbedDateNow("2026-03-11T00:10:00Z", () => {
-    const config = createConfig({
-      copilotReviewWaitMinutes: 10,
-      reviewBotLogins: ["copilot-pull-request-reviewer"],
-    });
-    const requestedAt = "2026-03-11T00:05:00Z";
-    const record = createRecord({
-      state: "waiting_ci",
-      review_wait_started_at: requestedAt,
-      review_wait_head_sha: "head123",
-    });
-    const pr: GitHubPullRequest = {
-      number: 44,
-      title: "Test PR",
-      url: "https://example.test/pr/44",
-      state: "OPEN",
-      createdAt: "2026-03-11T00:00:00Z",
-      isDraft: false,
-      reviewDecision: null,
-      mergeStateStatus: "CLEAN",
-      mergeable: "MERGEABLE",
-      headRefName: "codex/issue-38",
-      headRefOid: "head123",
-      mergedAt: null,
-      copilotReviewState: "requested",
-      copilotReviewRequestedAt: requestedAt,
-      copilotReviewArrivedAt: null,
-    };
-    const checks: PullRequestCheck[] = [{ name: "build", state: "SUCCESS", bucket: "pass", workflow: "CI" }];
-
-    assert.equal(inferStateFromPullRequest(config, record, pr, checks, []), "waiting_ci");
-  });
-});
-
-test("inferStateFromPullRequest treats an arrived configured-bot top-level review as satisfying the wait state", () => {
-  withStubbedDateNow("2026-03-11T00:10:00Z", () => {
-    const config = createConfig({
-      copilotReviewWaitMinutes: 10,
-      reviewBotLogins: ["coderabbitai[bot]"],
-    });
-    const requestedAt = "2026-03-11T00:05:00Z";
-    const record = createRecord({
-      state: "waiting_ci",
-      review_wait_started_at: requestedAt,
-      review_wait_head_sha: "head123",
-    });
-    const pr: GitHubPullRequest = {
-      number: 44,
-      title: "Test PR",
-      url: "https://example.test/pr/44",
-      state: "OPEN",
-      createdAt: "2026-03-11T00:00:00Z",
-      isDraft: false,
-      reviewDecision: null,
-      mergeStateStatus: "CLEAN",
-      mergeable: "MERGEABLE",
-      headRefName: "codex/issue-38",
-      headRefOid: "head123",
-      mergedAt: null,
-      copilotReviewState: "arrived",
-      copilotReviewRequestedAt: requestedAt,
-      copilotReviewArrivedAt: "2026-03-11T00:07:00Z",
-    };
-    const checks: PullRequestCheck[] = [{ name: "build", state: "SUCCESS", bucket: "pass", workflow: "CI" }];
-
-    assert.equal(inferStateFromPullRequest(config, record, pr, checks, []), "ready_to_merge");
-  });
-});
-
-test("inferStateFromPullRequest waits through a configured-bot rate limit warning for the configured window", () => {
-  withStubbedDateNow("2026-03-11T00:20:00Z", () => {
-    const config = createConfig({
-      reviewBotLogins: ["coderabbitai", "coderabbitai[bot]"],
-      configuredBotRateLimitWaitMinutes: 30,
-    });
-    const record = createRecord({ state: "waiting_ci" });
-    const pr: GitHubPullRequest = {
-      number: 44,
-      title: "Test PR",
-      url: "https://example.test/pr/44",
-      state: "OPEN",
-      createdAt: "2026-03-11T00:00:00Z",
-      isDraft: false,
-      reviewDecision: null,
-      mergeStateStatus: "CLEAN",
-      mergeable: "MERGEABLE",
-      headRefName: "codex/issue-38",
-      headRefOid: "head123",
-      mergedAt: null,
-      copilotReviewState: "requested",
-      copilotReviewRequestedAt: "2026-03-11T00:10:00Z",
-      copilotReviewArrivedAt: null,
-      configuredBotRateLimitedAt: "2026-03-11T00:15:00Z",
-    };
-    const checks: PullRequestCheck[] = [{ name: "build", state: "SUCCESS", bucket: "pass", workflow: "CI" }];
-
-    assert.equal(inferStateFromPullRequest(config, record, pr, checks, []), "waiting_ci");
-  });
-});
-
-test("inferStateFromPullRequest lets the rate-limit wait win over a blocking configured-bot timeout", () => {
-  withStubbedDateNow("2026-03-11T00:20:00Z", () => {
-    const config = createConfig({
-      reviewBotLogins: ["coderabbitai", "coderabbitai[bot]"],
-      configuredBotRateLimitWaitMinutes: 30,
-      copilotReviewWaitMinutes: 10,
-      copilotReviewTimeoutAction: "block",
-    });
-    const record = createRecord({
-      state: "waiting_ci",
-      copilot_review_requested_observed_at: "2026-03-11T00:00:00Z",
-      copilot_review_requested_head_sha: "head123",
-    });
-    const pr: GitHubPullRequest = {
-      number: 44,
-      title: "Test PR",
-      url: "https://example.test/pr/44",
-      state: "OPEN",
-      createdAt: "2026-03-11T00:00:00Z",
-      isDraft: false,
-      reviewDecision: null,
-      mergeStateStatus: "CLEAN",
-      mergeable: "MERGEABLE",
-      headRefName: "codex/issue-38",
-      headRefOid: "head123",
-      mergedAt: null,
-      copilotReviewState: "requested",
-      copilotReviewRequestedAt: "2026-03-11T00:00:00Z",
-      copilotReviewArrivedAt: null,
-      configuredBotRateLimitedAt: "2026-03-11T00:15:00Z",
-    };
-    const checks: PullRequestCheck[] = [{ name: "build", state: "SUCCESS", bucket: "pass", workflow: "CI" }];
-
-    assert.equal(inferStateFromPullRequest(config, record, pr, checks, []), "waiting_ci");
-  });
-});
-
-test("inferStateFromPullRequest allows merge again after a configured-bot rate limit wait expires", () => {
-  withStubbedDateNow("2026-03-11T00:50:01Z", () => {
-    const config = createConfig({
-      reviewBotLogins: ["coderabbitai", "coderabbitai[bot]"],
-      configuredBotRateLimitWaitMinutes: 30,
-    });
-    const record = createRecord({ state: "waiting_ci" });
-    const pr: GitHubPullRequest = {
-      number: 44,
-      title: "Test PR",
-      url: "https://example.test/pr/44",
-      state: "OPEN",
-      createdAt: "2026-03-11T00:00:00Z",
-      isDraft: false,
-      reviewDecision: null,
-      mergeStateStatus: "CLEAN",
-      mergeable: "MERGEABLE",
-      headRefName: "codex/issue-38",
-      headRefOid: "head123",
-      mergedAt: null,
-      copilotReviewState: "requested",
-      copilotReviewRequestedAt: "2026-03-11T00:10:00Z",
-      copilotReviewArrivedAt: null,
-      configuredBotRateLimitedAt: "2026-03-11T00:20:00Z",
-    };
-    const checks: PullRequestCheck[] = [{ name: "build", state: "SUCCESS", bucket: "pass", workflow: "CI" }];
-
-    assert.equal(inferStateFromPullRequest(config, record, pr, checks, []), "ready_to_merge");
-  });
-});
-
-test("inferStateFromPullRequest softens nitpick-only configured-bot top-level changes requests when no configured-bot threads remain", () => {
-  const config = createConfig({
-    reviewBotLogins: ["coderabbitai[bot]"],
-    humanReviewBlocksMerge: true,
-  });
-  const record = createRecord({ state: "pr_open" });
-  const pr = {
-    number: 44,
-    title: "Test PR",
-    url: "https://example.test/pr/44",
-    state: "OPEN",
-    createdAt: "2026-03-11T00:00:00Z",
-    isDraft: false,
-    reviewDecision: "CHANGES_REQUESTED",
-    mergeStateStatus: "CLEAN",
-    mergeable: "MERGEABLE",
-    headRefName: "codex/issue-38",
-    headRefOid: "head123",
-    mergedAt: null,
-    copilotReviewState: "arrived",
-    copilotReviewRequestedAt: "2026-03-11T00:05:00Z",
-    copilotReviewArrivedAt: "2026-03-11T00:07:00Z",
-    configuredBotTopLevelReviewStrength: "nitpick_only",
-  } as GitHubPullRequest;
-  const checks: PullRequestCheck[] = [{ name: "build", state: "SUCCESS", bucket: "pass", workflow: "CI" }];
-
-  assert.equal(inferStateFromPullRequest(config, record, pr, checks, []), "ready_to_merge");
-});
-
-test("inferStateFromPullRequest still blocks stronger configured-bot top-level changes requests without review threads", () => {
-  const config = createConfig({
-    reviewBotLogins: ["coderabbitai[bot]"],
-    humanReviewBlocksMerge: false,
-  });
-  const record = createRecord({ state: "pr_open" });
-  const pr: GitHubPullRequest = {
-    number: 44,
-    title: "Test PR",
-    url: "https://example.test/pr/44",
-    state: "OPEN",
-    createdAt: "2026-03-11T00:00:00Z",
-    isDraft: false,
-    reviewDecision: "CHANGES_REQUESTED",
-    mergeStateStatus: "CLEAN",
-    mergeable: "MERGEABLE",
-    headRefName: "codex/issue-38",
-    headRefOid: "head123",
-    mergedAt: null,
-    copilotReviewState: "arrived",
-    copilotReviewRequestedAt: "2026-03-11T00:05:00Z",
-    copilotReviewArrivedAt: "2026-03-11T00:07:00Z",
-    configuredBotTopLevelReviewStrength: "blocking",
-    configuredBotTopLevelReviewSubmittedAt: "2026-03-11T00:07:00Z",
-  };
-  const checks: PullRequestCheck[] = [{ name: "build", state: "SUCCESS", bucket: "pass", workflow: "CI" }];
-
-  assert.equal(inferStateFromPullRequest(config, record, pr, checks, []), "blocked");
-});
-
-test("inferStateFromPullRequest keeps waiting when a Copilot request was observed on the current head but has not arrived", () => {
-  withStubbedDateNow("2026-03-11T00:10:00Z", () => {
-    const config = createConfig({
-      copilotReviewWaitMinutes: 10,
-      reviewBotLogins: ["copilot-pull-request-reviewer"],
-    });
-    const record = createRecord({
-      state: "waiting_ci",
-      review_wait_started_at: "2026-03-11T00:00:00Z",
-      review_wait_head_sha: "head123",
-      copilot_review_requested_observed_at: "2026-03-11T00:05:00Z",
-      copilot_review_requested_head_sha: "head123",
-    });
-    const pr: GitHubPullRequest = {
-      number: 44,
-      title: "Test PR",
-      url: "https://example.test/pr/44",
-      state: "OPEN",
-      createdAt: "2026-03-11T00:00:00Z",
-      isDraft: false,
-      reviewDecision: null,
-      mergeStateStatus: "CLEAN",
-      mergeable: "MERGEABLE",
-      headRefName: "codex/issue-38",
-      headRefOid: "head123",
-      mergedAt: null,
-      copilotReviewState: "not_requested",
-      copilotReviewRequestedAt: null,
-      copilotReviewArrivedAt: null,
-    };
-    const checks: PullRequestCheck[] = [{ name: "build", state: "SUCCESS", bucket: "pass", workflow: "CI" }];
-
-    assert.equal(inferStateFromPullRequest(config, record, pr, checks, []), "waiting_ci");
-  });
-});
-
-test("inferStateFromPullRequest does not start Copilot timeout from the generic review wait window", () => {
-  withStubbedDateNow("2026-03-11T00:30:00Z", () => {
-    const config = createConfig({
-      copilotReviewWaitMinutes: 10,
-      copilotReviewTimeoutAction: "block",
-      reviewBotLogins: ["copilot-pull-request-reviewer"],
-    });
-    const record = createRecord({
-      state: "waiting_ci",
-      review_wait_started_at: "2026-03-11T00:00:00Z",
-      review_wait_head_sha: "head123",
-    });
-    const pr: GitHubPullRequest = {
-      number: 44,
-      title: "Test PR",
-      url: "https://example.test/pr/44",
-      state: "OPEN",
-      createdAt: "2026-03-11T00:00:00Z",
-      isDraft: false,
-      reviewDecision: null,
-      mergeStateStatus: "CLEAN",
-      mergeable: "MERGEABLE",
-      headRefName: "codex/issue-38",
-      headRefOid: "head123",
-      mergedAt: null,
-      copilotReviewState: "requested",
-      copilotReviewRequestedAt: null,
-      copilotReviewArrivedAt: null,
-    };
-    const checks: PullRequestCheck[] = [{ name: "build", state: "SUCCESS", bucket: "pass", workflow: "CI" }];
-
-    assert.equal(inferStateFromPullRequest(config, record, pr, checks, []), "waiting_ci");
-  });
-});
-
-test("inferStateFromPullRequest can time out from the observed Copilot request timestamp when GitHub omits one", () => {
-  withStubbedDateNow("2026-03-11T00:30:00Z", () => {
-    const config = createConfig({
-      copilotReviewWaitMinutes: 10,
-      copilotReviewTimeoutAction: "block",
-      reviewBotLogins: ["copilot-pull-request-reviewer"],
-    });
-    const record = createRecord({
-      state: "waiting_ci",
-      review_wait_started_at: "2026-03-11T00:00:00Z",
-      review_wait_head_sha: "head123",
-      copilot_review_requested_observed_at: "2026-03-11T00:15:00Z",
-      copilot_review_requested_head_sha: "head123",
-    });
-    const pr: GitHubPullRequest = {
-      number: 44,
-      title: "Test PR",
-      url: "https://example.test/pr/44",
-      state: "OPEN",
-      createdAt: "2026-03-11T00:00:00Z",
-      isDraft: false,
-      reviewDecision: null,
-      mergeStateStatus: "CLEAN",
-      mergeable: "MERGEABLE",
-      headRefName: "codex/issue-38",
-      headRefOid: "head123",
-      mergedAt: null,
-      copilotReviewState: "requested",
-      copilotReviewRequestedAt: null,
-      copilotReviewArrivedAt: null,
-    };
-    const checks: PullRequestCheck[] = [{ name: "build", state: "SUCCESS", bucket: "pass", workflow: "CI" }];
-
-    assert.equal(inferStateFromPullRequest(config, record, pr, checks, []), "blocked");
-  });
-});
-
-test("inferStateFromPullRequest can block when a configured bot review times out from the observed request timestamp", () => {
-  withStubbedDateNow("2026-03-11T00:30:00Z", () => {
-    const config = createConfig({
-      copilotReviewWaitMinutes: 10,
-      copilotReviewTimeoutAction: "block",
-      reviewBotLogins: ["chatgpt-codex-connector"],
-    });
-    const record = createRecord({
-      state: "waiting_ci",
-      review_wait_started_at: "2026-03-11T00:00:00Z",
-      review_wait_head_sha: "head123",
-      copilot_review_requested_observed_at: "2026-03-11T00:15:00Z",
-      copilot_review_requested_head_sha: "head123",
-    });
-    const pr: GitHubPullRequest = {
-      number: 44,
-      title: "Test PR",
-      url: "https://example.test/pr/44",
-      state: "OPEN",
-      createdAt: "2026-03-11T00:00:00Z",
-      isDraft: false,
-      reviewDecision: null,
-      mergeStateStatus: "CLEAN",
-      mergeable: "MERGEABLE",
-      headRefName: "codex/issue-38",
-      headRefOid: "head123",
-      mergedAt: null,
-      copilotReviewState: "requested",
-      copilotReviewRequestedAt: null,
-      copilotReviewArrivedAt: null,
-    };
-    const checks: PullRequestCheck[] = [{ name: "build", state: "SUCCESS", bucket: "pass", workflow: "CI" }];
-
-    assert.equal(inferStateFromPullRequest(config, record, pr, checks, []), "blocked");
-  });
-});
-
-test("inferStateFromPullRequest does not wait on stale configured bot request state when no review bots are configured", () => {
-  withStubbedDateNow("2026-03-11T00:30:00Z", () => {
-    const config = createConfig({
-      copilotReviewWaitMinutes: 10,
-      copilotReviewTimeoutAction: "block",
-      reviewBotLogins: [],
-    });
-    const record = createRecord({
-      state: "waiting_ci",
-      review_wait_started_at: "2026-03-11T00:00:00Z",
-      review_wait_head_sha: "head123",
-      copilot_review_requested_observed_at: "2026-03-11T00:05:00Z",
-      copilot_review_requested_head_sha: "head123",
-    });
-    const pr: GitHubPullRequest = {
-      number: 44,
-      title: "Test PR",
-      url: "https://example.test/pr/44",
-      state: "OPEN",
-      createdAt: "2026-03-11T00:00:00Z",
-      isDraft: false,
-      reviewDecision: null,
-      mergeStateStatus: "CLEAN",
-      mergeable: "MERGEABLE",
-      headRefName: "codex/issue-38",
-      headRefOid: "head123",
-      mergedAt: null,
-      copilotReviewState: "not_requested",
-      copilotReviewRequestedAt: null,
-      copilotReviewArrivedAt: null,
-    };
-    const checks: PullRequestCheck[] = [{ name: "build", state: "SUCCESS", bucket: "pass", workflow: "CI" }];
-
-    assert.equal(inferStateFromPullRequest(config, record, pr, checks, []), "ready_to_merge");
-  });
-});
-
-test("inferStateFromPullRequest times out requested Copilot reviews and continues by default", () => {
-  withStubbedDateNow("2026-03-11T00:30:00Z", () => {
-    const config = createConfig({ copilotReviewWaitMinutes: 10, copilotReviewTimeoutAction: "continue" });
-    const record = createRecord({
-      state: "waiting_ci",
-      review_wait_started_at: "2026-03-11T00:00:00Z",
-      review_wait_head_sha: "head123",
-    });
-    const pr: GitHubPullRequest = {
-      number: 44,
-      title: "Test PR",
-      url: "https://example.test/pr/44",
-      state: "OPEN",
-      createdAt: "2026-03-11T00:00:00Z",
-      isDraft: false,
-      reviewDecision: null,
-      mergeStateStatus: "CLEAN",
-      mergeable: "MERGEABLE",
-      headRefName: "codex/issue-38",
-      headRefOid: "head123",
-      mergedAt: null,
-      copilotReviewState: "requested",
-      copilotReviewRequestedAt: "2026-03-11T00:05:00Z",
-      copilotReviewArrivedAt: null,
-    };
-    const checks: PullRequestCheck[] = [{ name: "build", state: "SUCCESS", bucket: "pass", workflow: "CI" }];
-
-    assert.equal(inferStateFromPullRequest(config, record, pr, checks, []), "ready_to_merge");
-  });
-});
-
-test("inferStateFromPullRequest can block when a requested Copilot review times out", () => {
-  withStubbedDateNow("2026-03-11T00:30:00Z", () => {
-    const config = createConfig({
-      copilotReviewWaitMinutes: 10,
-      copilotReviewTimeoutAction: "block",
-      reviewBotLogins: ["copilot-pull-request-reviewer"],
-    });
-    const record = createRecord({
-      state: "waiting_ci",
-      review_wait_started_at: "2026-03-11T00:00:00Z",
-      review_wait_head_sha: "head123",
-    });
-    const pr: GitHubPullRequest = {
-      number: 44,
-      title: "Test PR",
-      url: "https://example.test/pr/44",
-      state: "OPEN",
-      createdAt: "2026-03-11T00:00:00Z",
-      isDraft: false,
-      reviewDecision: null,
-      mergeStateStatus: "CLEAN",
-      mergeable: "MERGEABLE",
-      headRefName: "codex/issue-38",
-      headRefOid: "head123",
-      mergedAt: null,
-      copilotReviewState: "requested",
-      copilotReviewRequestedAt: "2026-03-11T00:05:00Z",
-      copilotReviewArrivedAt: null,
-    };
-    const checks: PullRequestCheck[] = [{ name: "build", state: "SUCCESS", bucket: "pass", workflow: "CI" }];
-
-    assert.equal(inferStateFromPullRequest(config, record, pr, checks, []), "blocked");
-  });
-});
-
-test("inferStateFromPullRequest covers local review policy gating combinations", () => {
-  const cases: Array<{
-    name: string;
-    config: Partial<SupervisorConfig>;
-    record: Partial<IssueRunRecord>;
-    pr: Partial<GitHubPullRequest>;
-    expected: IssueRunRecord["state"];
-  }> = [
-    {
-      name: "block_ready keeps draft PRs in draft_pr when raw findings exist on the current head",
-      config: { localReviewEnabled: true, localReviewPolicy: "block_ready", copilotReviewWaitMinutes: 0 },
-      record: {
-        state: "draft_pr",
-        local_review_head_sha: "head123",
-        local_review_findings_count: 2,
-        local_review_recommendation: "changes_requested",
-      },
-      pr: { isDraft: true, headRefOid: "head123" },
-      expected: "draft_pr",
-    },
-    {
-      name: "block_ready does not block a ready PR after it becomes ready",
-      config: { localReviewEnabled: true, localReviewPolicy: "block_ready", copilotReviewWaitMinutes: 0 },
-      record: {
-        state: "pr_open",
-        local_review_head_sha: "head123",
-        local_review_findings_count: 2,
-        local_review_recommendation: "changes_requested",
-      },
-      pr: { isDraft: false, headRefOid: "head123" },
-      expected: "ready_to_merge",
-    },
-    {
-      name: "block_merge blocks merge for ready PRs with actionable findings on the current head",
-      config: { localReviewEnabled: true, localReviewPolicy: "block_merge", copilotReviewWaitMinutes: 0 },
-      record: {
-        state: "pr_open",
-        local_review_head_sha: "head123",
-        local_review_findings_count: 2,
-        local_review_recommendation: "changes_requested",
-      },
-      pr: { isDraft: false, headRefOid: "head123" },
-      expected: "blocked",
-    },
-    {
-      name: "block_merge stops gating once the review head becomes stale",
-      config: { localReviewEnabled: true, localReviewPolicy: "block_merge", copilotReviewWaitMinutes: 0 },
-      record: {
-        state: "pr_open",
-        local_review_head_sha: "oldhead",
-        local_review_findings_count: 2,
-        local_review_recommendation: "changes_requested",
-      },
-      pr: { isDraft: false, headRefOid: "newhead" },
-      expected: "ready_to_merge",
-    },
-    {
-      name: "advisory never blocks merge for ready PRs with raw findings",
-      config: { localReviewEnabled: true, localReviewPolicy: "advisory", copilotReviewWaitMinutes: 0 },
-      record: {
-        state: "pr_open",
-        local_review_head_sha: "head123",
-        local_review_findings_count: 2,
-        local_review_recommendation: "changes_requested",
-      },
-      pr: { isDraft: false, headRefOid: "head123" },
-      expected: "ready_to_merge",
-    },
-    {
-      name: "retry escalates verifier-confirmed high severity findings into local_review_fix",
-      config: {
-        localReviewEnabled: true,
-        localReviewPolicy: "block_merge",
-        localReviewHighSeverityAction: "retry",
-        copilotReviewWaitMinutes: 0,
-      },
-      record: {
-        state: "pr_open",
-        local_review_head_sha: "head123",
-        local_review_findings_count: 3,
-        local_review_recommendation: "changes_requested",
-        local_review_verified_max_severity: "high",
-        local_review_verified_findings_count: 1,
-        repeated_local_review_signature_count: 1,
-      },
-      pr: { isDraft: false, headRefOid: "head123" },
-      expected: "local_review_fix",
-    },
-    {
-      name: "blocked escalates verifier-confirmed high severity findings to blocked",
-      config: {
-        localReviewEnabled: true,
-        localReviewPolicy: "block_merge",
-        localReviewHighSeverityAction: "blocked",
-        copilotReviewWaitMinutes: 0,
-      },
-      record: {
-        state: "pr_open",
-        local_review_head_sha: "head123",
-        local_review_findings_count: 3,
-        local_review_recommendation: "changes_requested",
-        local_review_verified_max_severity: "high",
-        local_review_verified_findings_count: 1,
-      },
-      pr: { isDraft: false, headRefOid: "head123" },
-      expected: "blocked",
-    },
-    {
-      name: "advisory suppresses high severity retry escalation",
-      config: {
-        localReviewEnabled: true,
-        localReviewPolicy: "advisory",
-        localReviewHighSeverityAction: "retry",
-        copilotReviewWaitMinutes: 0,
-      },
-      record: {
-        state: "pr_open",
-        local_review_head_sha: "head123",
-        local_review_findings_count: 3,
-        local_review_recommendation: "changes_requested",
-        local_review_verified_max_severity: "high",
-        local_review_verified_findings_count: 1,
-      },
-      pr: { isDraft: false, headRefOid: "head123" },
-      expected: "ready_to_merge",
-    },
-  ];
-
-  for (const testCase of cases) {
-    const config = createConfig(testCase.config);
-    const record = createRecord(testCase.record);
-    const pr: GitHubPullRequest = {
-      number: 44,
-      title: "Test PR",
-      url: "https://example.test/pr/44",
-      state: "OPEN",
-      createdAt: "2026-03-01T00:00:00Z",
-      isDraft: false,
-      reviewDecision: null,
-      mergeStateStatus: "CLEAN",
-      mergeable: "MERGEABLE",
-      headRefName: "codex/issue-38",
-      headRefOid: "head123",
-      mergedAt: null,
-      ...testCase.pr,
-    };
-
-    assert.equal(
-      inferStateFromPullRequest(config, record, pr, [], []),
-      testCase.expected,
-      testCase.name,
-    );
-  }
-});
-
-test("inferStateFromPullRequest blocks stalled identical high local-review retries", () => {
-  const config = createConfig({
-    localReviewEnabled: true,
-    localReviewPolicy: "block_ready",
-    localReviewHighSeverityAction: "retry",
-    sameFailureSignatureRepeatLimit: 3,
-  });
-  const record = createRecord({
-    state: "draft_pr",
-    pr_number: 44,
-    local_review_head_sha: "head123",
-    local_review_max_severity: "high",
-    local_review_verified_max_severity: "high",
-    local_review_verified_findings_count: 1,
-    local_review_findings_count: 3,
-    local_review_recommendation: "changes_requested",
-    repeated_local_review_signature_count: 3,
-  });
-  const pr: GitHubPullRequest = {
-    number: 44,
-    title: "Test PR",
-    url: "https://example.test/pr/44",
-    state: "OPEN",
-    createdAt: "2026-03-11T00:00:00Z",
-    isDraft: true,
-    reviewDecision: null,
-    mergeStateStatus: "CLEAN",
-    headRefName: "codex/issue-38",
-    headRefOid: "head123",
-  };
-
-  assert.equal(inferStateFromPullRequest(config, record, pr, [], []), "blocked");
-});
-
-test("inferStateFromPullRequest does not stall local-review retries when CI adds a fresh signal", () => {
-  const config = createConfig({
-    localReviewEnabled: true,
-    localReviewPolicy: "block_ready",
-    localReviewHighSeverityAction: "retry",
-    sameFailureSignatureRepeatLimit: 3,
-  });
-  const record = createRecord({
-    state: "local_review_fix",
-    pr_number: 44,
-    local_review_head_sha: "head123",
-    local_review_max_severity: "high",
-    local_review_verified_max_severity: "high",
-    local_review_verified_findings_count: 1,
-    local_review_findings_count: 3,
-    local_review_recommendation: "changes_requested",
-    repeated_local_review_signature_count: 3,
-  });
-  const pr: GitHubPullRequest = {
-    number: 44,
-    title: "Test PR",
-    url: "https://example.test/pr/44",
-    state: "OPEN",
-    createdAt: "2026-03-11T00:00:00Z",
-    isDraft: true,
-    reviewDecision: null,
-    mergeStateStatus: "CLEAN",
-    headRefName: "codex/issue-38",
-    headRefOid: "head123",
-  };
-  const checks: PullRequestCheck[] = [{ name: "test", state: "FAILURE", bucket: "fail", workflow: "CI" }];
-
-  assert.equal(inferStateFromPullRequest(config, record, pr, checks, []), "local_review_fix");
-});
-
-test("inferStateFromPullRequest keeps an unresolved configured bot thread blocked on the same head after processing", () => {
-  const config = createConfig({
-    reviewBotLogins: ["copilot-pull-request-reviewer"],
-  });
-  const record = createRecord({
-    state: "pr_open",
-    last_head_sha: "head-a",
-    processed_review_thread_ids: ["thread-1@head-a"],
-    processed_review_thread_fingerprints: ["thread-1@head-a#comment-1"],
-  });
-  const pr: GitHubPullRequest = {
-    number: 44,
-    title: "Test PR",
-    url: "https://example.test/pr/44",
-    state: "OPEN",
-    createdAt: "2026-03-11T00:00:00Z",
-    isDraft: false,
-    reviewDecision: "CHANGES_REQUESTED",
-    mergeStateStatus: "CLEAN",
-    mergeable: "MERGEABLE",
-    headRefName: "codex/issue-38",
-    headRefOid: "head-a",
-    mergedAt: null,
-  };
-
-  assert.equal(
-    inferStateFromPullRequest(config, record, pr, [], [createReviewThread()]),
-    "blocked",
-  );
-});
-
-test("inferStateFromPullRequest treats a legacy plain thread id as processed only on the matching head", () => {
-  const config = createConfig({
-    reviewBotLogins: ["copilot-pull-request-reviewer"],
-  });
-  const record = createRecord({
-    state: "pr_open",
-    last_head_sha: "head-a",
-    processed_review_thread_ids: ["thread-1"],
-  });
-  const sameHeadPr: GitHubPullRequest = {
-    number: 44,
-    title: "Test PR",
-    url: "https://example.test/pr/44",
-    state: "OPEN",
-    createdAt: "2026-03-11T00:00:00Z",
-    isDraft: false,
-    reviewDecision: "CHANGES_REQUESTED",
-    mergeStateStatus: "CLEAN",
-    mergeable: "MERGEABLE",
-    headRefName: "codex/issue-38",
-    headRefOid: "head-a",
-    mergedAt: null,
-  };
-  const changedHeadPr: GitHubPullRequest = {
-    ...sameHeadPr,
-    headRefOid: "head-b",
-  };
-
-  assert.equal(
-    inferStateFromPullRequest(config, record, sameHeadPr, [], [createReviewThread()]),
-    "blocked",
-  );
-  assert.equal(
-    inferStateFromPullRequest(config, record, changedHeadPr, [], [createReviewThread()]),
-    "addressing_review",
-  );
-});
-
-test("inferStateFromPullRequest allows one reprocessing pass for a configured bot thread after the PR head changes", () => {
-  const config = createConfig({
-    reviewBotLogins: ["copilot-pull-request-reviewer"],
-  });
-  const record = createRecord({
-    state: "pr_open",
-    last_head_sha: "head-a",
-    processed_review_thread_ids: ["thread-1@head-a"],
-    processed_review_thread_fingerprints: ["thread-1@head-a#comment-1"],
-  });
-  const pr: GitHubPullRequest = {
-    number: 44,
-    title: "Test PR",
-    url: "https://example.test/pr/44",
-    state: "OPEN",
-    createdAt: "2026-03-11T00:00:00Z",
-    isDraft: false,
-    reviewDecision: "CHANGES_REQUESTED",
-    mergeStateStatus: "CLEAN",
-    mergeable: "MERGEABLE",
-    headRefName: "codex/issue-38",
-    headRefOid: "head-b",
-    mergedAt: null,
-  };
-
-  assert.equal(
-    inferStateFromPullRequest(config, record, pr, [], [createReviewThread()]),
-    "addressing_review",
-  );
-});
-
-test("inferStateFromPullRequest allows one reprocessing pass for a configured bot thread when its latest comment changes on the same head", () => {
-  const config = createConfig({
-    reviewBotLogins: ["copilot-pull-request-reviewer"],
-  });
-  const record = createRecord({
-    state: "pr_open",
-    last_head_sha: "head-a",
-    processed_review_thread_ids: ["thread-1@head-a"],
-    processed_review_thread_fingerprints: ["thread-1@head-a#comment-1"],
-  });
-  const pr: GitHubPullRequest = {
-    number: 44,
-    title: "Test PR",
-    url: "https://example.test/pr/44",
-    state: "OPEN",
-    createdAt: "2026-03-11T00:00:00Z",
-    isDraft: false,
-    reviewDecision: "CHANGES_REQUESTED",
-    mergeStateStatus: "CLEAN",
-    mergeable: "MERGEABLE",
-    headRefName: "codex/issue-38",
-    headRefOid: "head-a",
-    mergedAt: null,
-  };
-  const updatedThread = createReviewThread({
-    comments: {
-      nodes: [
-        {
-          id: "comment-1",
-          body: "Please address this.",
-          createdAt: "2026-03-11T00:00:00Z",
-          url: "https://example.test/pr/44#discussion_r1",
-          author: {
-            login: "copilot-pull-request-reviewer",
-            typeName: "Bot",
-          },
-        },
-        {
-          id: "comment-2",
-          body: "One more note on the same thread.",
-          createdAt: "2026-03-11T00:05:00Z",
-          url: "https://example.test/pr/44#discussion_r2",
-          author: {
-            login: "copilot-pull-request-reviewer",
-            typeName: "Bot",
-          },
-        },
-      ],
-    },
-  });
-
-  assert.equal(
-    inferStateFromPullRequest(config, record, pr, [], [updatedThread]),
-    "addressing_review",
-  );
-});
-
-test("inferStateFromPullRequest blocks a same-head configured bot thread again after its updated comment has already been reprocessed once", () => {
-  const config = createConfig({
-    reviewBotLogins: ["copilot-pull-request-reviewer"],
-  });
-  const record = createRecord({
-    state: "pr_open",
-    last_head_sha: "head-a",
-    processed_review_thread_ids: ["thread-1@head-a"],
-    processed_review_thread_fingerprints: ["thread-1@head-a#comment-2"],
-  });
-  const pr: GitHubPullRequest = {
-    number: 44,
-    title: "Test PR",
-    url: "https://example.test/pr/44",
-    state: "OPEN",
-    createdAt: "2026-03-11T00:00:00Z",
-    isDraft: false,
-    reviewDecision: "CHANGES_REQUESTED",
-    mergeStateStatus: "CLEAN",
-    mergeable: "MERGEABLE",
-    headRefName: "codex/issue-38",
-    headRefOid: "head-a",
-    mergedAt: null,
-  };
-  const updatedThread = createReviewThread({
-    comments: {
-      nodes: [
-        {
-          id: "comment-1",
-          body: "Please address this.",
-          createdAt: "2026-03-11T00:00:00Z",
-          url: "https://example.test/pr/44#discussion_r1",
-          author: {
-            login: "copilot-pull-request-reviewer",
-            typeName: "Bot",
-          },
-        },
-        {
-          id: "comment-2",
-          body: "One more note on the same thread.",
-          createdAt: "2026-03-11T00:05:00Z",
-          url: "https://example.test/pr/44#discussion_r2",
-          author: {
-            login: "copilot-pull-request-reviewer",
-            typeName: "Bot",
-          },
-        },
-      ],
-    },
-  });
-
-  assert.equal(
-    inferStateFromPullRequest(config, record, pr, [], [updatedThread]),
-    "blocked",
-  );
-});
-
-test("inferStateFromPullRequest blocks a repeatedly unresolved configured bot thread again after its one pass on the new head", () => {
-  const config = createConfig({
-    reviewBotLogins: ["copilot-pull-request-reviewer"],
-  });
-  const record = createRecord({
-    state: "pr_open",
-    last_head_sha: "head-b",
-    processed_review_thread_ids: ["thread-1@head-a", "thread-1@head-b"],
-    processed_review_thread_fingerprints: ["thread-1@head-b#comment-1"],
-  });
-  const pr: GitHubPullRequest = {
-    number: 44,
-    title: "Test PR",
-    url: "https://example.test/pr/44",
-    state: "OPEN",
-    createdAt: "2026-03-11T00:00:00Z",
-    isDraft: false,
-    reviewDecision: "CHANGES_REQUESTED",
-    mergeStateStatus: "CLEAN",
-    mergeable: "MERGEABLE",
-    headRefName: "codex/issue-38",
-    headRefOid: "head-b",
-    mergedAt: null,
-  };
-
-  assert.equal(
-    inferStateFromPullRequest(config, record, pr, [], [createReviewThread()]),
-    "blocked",
-  );
-});
-
-
-test("formatDetailedStatus shows blocking local review status for current PR head", () => {
-  const config = createConfig({ localReviewPolicy: "block_ready" });
-  const record = createRecord({
-    local_review_head_sha: "deadbeef",
-    local_review_blocker_summary: "high src/supervisor.ts:210-214 stale artifact context drives the wrong repair path.",
-    local_review_max_severity: "high",
-    local_review_findings_count: 3,
-    local_review_recommendation: "changes_requested",
-    local_review_run_at: "2026-03-11T14:05:00Z",
-  });
-  const pr: GitHubPullRequest = {
-    number: 42,
-    title: "Test PR",
-    url: "https://example.test/pr/42",
-    state: "OPEN",
-    createdAt: "2026-03-11T14:00:00Z",
-    isDraft: true,
-    reviewDecision: null,
-    mergeStateStatus: "CLEAN",
-    mergeable: "MERGEABLE",
-    headRefName: "codex/issue-42",
-    headRefOid: "deadbeef",
-    mergedAt: null,
-  };
-
-  const status = formatDetailedStatus({
-    config,
-    activeRecord: record,
-    latestRecord: record,
-    trackedIssueCount: 1,
-    pr,
-    checks: [],
-    reviewThreads: [],
-  });
-
-  assert.match(
-    status,
-    /local_review gating=yes policy=block_ready findings=3 root_causes=0 max_severity=high verified_findings=0 verified_max_severity=none head=current reviewed_head_sha=deadbeef pr_head_sha=deadbeef ran_at=2026-03-11T14:05:00Z blocker_summary=high src\/supervisor\.ts:210-214 stale artifact context drives the wrong repair path\. signature=none repeated=0 stalled=no/,
-  );
-  assert.doesNotMatch(status, /needs_review_run=/);
-  assert.match(status, /external_review head=none reviewed_head_sha=none matched=0 near_match=0 missed=0/);
-});
-
-test("buildDetailedStatusModel returns the reusable core status lines for an active PR", () => {
-  const config = createConfig({
-    localReviewPolicy: "block_ready",
-    reviewBotLogins: ["chatgpt-codex-connector"],
-  });
-  const record = createRecord({
-    local_review_head_sha: "deadbeef",
-    local_review_blocker_summary: "high src/supervisor.ts:210-214 stale artifact context drives the wrong repair path.",
-    local_review_max_severity: "high",
-    local_review_findings_count: 3,
-    local_review_recommendation: "changes_requested",
-    local_review_run_at: "2026-03-11T14:05:00Z",
-    pr_number: 44,
-    state: "pr_open",
-    blocked_reason: null,
-    last_error: null,
-    last_failure_context: null,
-    external_review_head_sha: null,
-    external_review_matched_findings_count: 0,
-    external_review_near_match_findings_count: 0,
-    external_review_missed_findings_count: 0,
-  });
-  const pr: GitHubPullRequest = {
-    number: 44,
-    title: "Test PR",
-    url: "https://example.test/pr/44",
-    state: "OPEN",
-    createdAt: "2026-03-11T14:00:00Z",
-    isDraft: false,
-    reviewDecision: "REVIEW_REQUIRED",
-    mergeStateStatus: "CLEAN",
-    mergeable: "MERGEABLE",
-    headRefName: "codex/issue-44",
-    headRefOid: "deadbeef",
-    mergedAt: null,
-    copilotReviewState: "not_requested",
-    copilotReviewRequestedAt: null,
-    copilotReviewArrivedAt: null,
-  };
-
-  const lines = buildDetailedStatusModel({
-    config,
-    activeRecord: record,
-    latestRecord: null,
-    trackedIssueCount: 1,
-    pr,
-    checks: [{ name: "build", state: "SUCCESS", bucket: "pass", workflow: "CI" }],
-    reviewThreads: [],
-    manualReviewThreads,
-    configuredBotReviewThreads,
-    pendingBotReviewThreads: (innerConfig, innerRecord, innerPr, innerReviewThreads) =>
-      configuredBotReviewThreads(innerConfig, innerReviewThreads).filter(
-        (thread) =>
-          !innerRecord.processed_review_thread_ids.includes(thread.id) &&
-          innerRecord.last_head_sha === innerPr.headRefOid,
-      ),
-    summarizeChecks,
-    mergeConflictDetected: (innerPr) => innerPr.mergeStateStatus === "DIRTY",
-  });
-
-  assert.ok(lines.includes("issue=#366"));
-  assert.ok(
-    lines.some((line) =>
-      /local_review gating=yes policy=block_ready findings=3 .* head=current .* blocker_summary=high src\/supervisor\.ts:210-214 stale artifact context drives the wrong repair path\./.test(
-        line,
-      ),
-    ),
-  );
-  assert.ok(lines.includes("external_review head=none reviewed_head_sha=none matched=0 near_match=0 missed=0"));
-  assert.ok(
-    lines.includes(
-      "review_bot_profile profile=codex provider=chatgpt-codex-connector reviewers=chatgpt-codex-connector signal_source=review_threads",
-    ),
-  );
-  assert.ok(
-    lines.includes(
-      "review_bot_diagnostics status=missing_provider_signal observed_review=none expected_reviewers=chatgpt-codex-connector next_check=provider_setup_or_delivery",
-    ),
-  );
-});
-
-test("buildDetailedStatusSummaryLines shapes optional summaries and artifact paths", () => {
-  const config = createConfig({
-    localReviewArtifactDir: "/tmp/reviews",
-  });
-  const activeRecord = createRecord({
-    local_review_summary_path: "/tmp/reviews/owner-repo/issue-58/local-review-summary.md",
-    external_review_misses_path: "/tmp/reviews/owner-repo/issue-58/external-review-misses-head-deadbeef.json",
-  });
-  const latestRecoveryRecord = createRecord({
-    issue_number: 91,
-    state: "done",
-    branch: "codex/issue-91",
-    workspace: "/tmp/workspaces/issue-91",
-    updated_at: "2026-03-13T00:20:00Z",
-    last_codex_summary: null,
-    last_recovery_reason: "merged_pr_convergence: tracked PR #191 merged; marked issue #91 done",
-    last_recovery_at: "2026-03-13T00:20:00Z",
-  });
-
-  assert.deepEqual(
-    buildDetailedStatusSummaryLines({
-      config,
-      activeRecord,
-      latestRecoveryRecord,
-      handoffSummary: "blocked\nneeds reproduction",
-      durableGuardrailSummary: "durable_guardrails verifier=committed:.codex/verifier-guardrails.json#1 external_review=none",
-    }),
-    [
-      "handoff_summary=blocked\\nneeds reproduction",
-      "durable_guardrails verifier=committed:.codex/verifier-guardrails.json#1 external_review=none",
-      "latest_recovery issue=#91 at=2026-03-13T00:20:00Z reason=merged_pr_convergence: tracked PR #191 merged; marked issue #91 done",
-      "local_review_summary_path=owner-repo/issue-58/local-review-summary.md",
-      "external_review_misses_path=owner-repo/issue-58/external-review-misses-head-deadbeef.json",
-    ],
-  );
-});
-
-test("buildDetailedStatusModel sanitizes failure context summary before emitting it", () => {
-  const lines = buildDetailedStatusModel({
-    config: createConfig(),
-    activeRecord: createRecord({
-      last_error: null,
-      last_failure_context: {
-        category: "blocked",
-        summary: "first line\nsecond line",
-        signature: "two-line-summary",
-        command: null,
-        details: [],
-        url: null,
-        updated_at: "2026-03-11T01:50:41.997Z",
-      },
-    }),
-    latestRecord: null,
-    trackedIssueCount: 1,
-    pr: null,
-    checks: [],
-    reviewThreads: [],
-    manualReviewThreads,
-    configuredBotReviewThreads,
-    pendingBotReviewThreads: () => [],
-    summarizeChecks,
-    mergeConflictDetected: (innerPr) => innerPr.mergeStateStatus === "DIRTY",
-  });
-
-  assert.ok(lines.includes("failure_context category=blocked summary=first line\\nsecond line"));
-});
-
-test("buildDetailedStatusModel counts only unresolved configured bot threads in status fields", () => {
-  const config = createConfig({
-    reviewBotLogins: ["coderabbitai[bot]"],
-  });
-  const pr: GitHubPullRequest = {
-    number: 44,
-    title: "Test PR",
-    url: "https://example.test/pr/44",
-    state: "OPEN",
-    createdAt: "2026-03-11T14:00:00Z",
-    isDraft: false,
-    reviewDecision: "CHANGES_REQUESTED",
-    mergeStateStatus: "CLEAN",
-    mergeable: "MERGEABLE",
-    headRefName: "codex/issue-38",
-    headRefOid: "deadbeef",
-    mergedAt: null,
-    copilotReviewState: "arrived",
-    copilotReviewRequestedAt: "2026-03-11T14:05:00Z",
-    copilotReviewArrivedAt: "2026-03-11T14:06:00Z",
-    configuredBotTopLevelReviewStrength: "nitpick_only",
-    configuredBotTopLevelReviewSubmittedAt: "2026-03-11T14:06:00Z",
-  };
-  const resolvedBotThread = createReviewThread({
-    isResolved: true,
-    comments: {
-      nodes: [
-        {
-          id: "comment-1",
-          body: "Resolved.",
-          createdAt: "2026-03-11T00:00:00Z",
-          url: "https://example.test/pr/44#discussion_r1",
-          author: {
-            login: "coderabbitai[bot]",
-            typeName: "Bot",
-          },
-        },
-      ],
-    },
-  });
-
-  const lines = buildDetailedStatusModel({
-    config,
-    activeRecord: createRecord({
-      pr_number: 44,
-      state: "ready_to_merge",
-      last_error: null,
-      last_failure_context: null,
-    }),
-    latestRecord: null,
-    trackedIssueCount: 1,
-    pr,
-    checks: [{ name: "build", state: "SUCCESS", bucket: "pass", workflow: "CI" }],
-    reviewThreads: [resolvedBotThread],
-    manualReviewThreads,
-    configuredBotReviewThreads,
-    pendingBotReviewThreads: () => [],
-    summarizeChecks,
-    mergeConflictDetected: (innerPr) => innerPr.mergeStateStatus === "DIRTY",
-  });
-
-  assert.ok(
-    lines.includes(
-      "configured_bot_top_level_review strength=nitpick_only submitted_at=2026-03-11T14:06:00Z effect=softened",
-    ),
-  );
-  assert.ok(lines.includes("review_threads bot_pending=0 bot_unresolved=0 manual=0"));
-});
-
-test("formatDetailedStatus shows both raw and compressed local review counts", () => {
-  const config = createConfig({ localReviewPolicy: "block_ready" });
-  const record = {
-    ...createRecord({
-      local_review_head_sha: "deadbeef",
-      local_review_max_severity: "high",
-      local_review_findings_count: 3,
-      local_review_root_cause_count: 1,
-      local_review_recommendation: "changes_requested",
-      local_review_run_at: "2026-03-11T14:05:00Z",
-    }),
-  };
-  const pr: GitHubPullRequest = {
-    number: 42,
-    title: "Test PR",
-    url: "https://example.test/pr/42",
-    state: "OPEN",
-    createdAt: "2026-03-11T14:00:00Z",
-    isDraft: true,
-    reviewDecision: null,
-    mergeStateStatus: "CLEAN",
-    mergeable: "MERGEABLE",
-    headRefName: "codex/issue-42",
-    headRefOid: "deadbeef",
-    mergedAt: null,
-  };
-
-  const status = formatDetailedStatus({
-    config,
-    activeRecord: record,
-    latestRecord: record,
-    trackedIssueCount: 1,
-    pr,
-    checks: [],
-    reviewThreads: [],
-  });
-
-  assert.match(status, /local_review .*findings=3 .*root_causes=1 .*stalled=no/);
-});
-
-test("formatDetailedStatus marks stalled local-review repair loops explicitly", () => {
-  const config = createConfig({
-    localReviewEnabled: true,
-    localReviewPolicy: "block_ready",
-    localReviewHighSeverityAction: "retry",
-    sameFailureSignatureRepeatLimit: 3,
-  });
-  const record = createRecord({
-    state: "blocked",
-    blocked_reason: "verification",
-    local_review_head_sha: "deadbeef",
-    local_review_max_severity: "high",
-    local_review_verified_max_severity: "high",
-    local_review_verified_findings_count: 1,
-    local_review_findings_count: 3,
-    local_review_root_cause_count: 1,
-    local_review_recommendation: "changes_requested",
-    repeated_local_review_signature_count: 3,
-  });
-  const pr: GitHubPullRequest = {
-    number: 42,
-    title: "Test PR",
-    url: "https://example.test/pr/42",
-    state: "OPEN",
-    createdAt: "2026-03-11T14:00:00Z",
-    isDraft: true,
-    reviewDecision: null,
-    mergeStateStatus: "CLEAN",
-    mergeable: "MERGEABLE",
-    headRefName: "codex/issue-42",
-    headRefOid: "deadbeef",
-    mergedAt: null,
-  };
-
-  const status = formatDetailedStatus({
-    config,
-    activeRecord: record,
-    latestRecord: record,
-    trackedIssueCount: 1,
-    pr,
-    checks: [],
-    reviewThreads: [],
-  });
-
-  assert.match(status, /local_review .* repeated=3 stalled=yes/);
-  assert.match(status, /blocked_reason=verification/);
-});
-
-test("formatDetailedStatus shows saved external review miss counts for the current PR head", () => {
-  const config = createConfig();
-  const pr: GitHubPullRequest = {
-    number: 22,
-    title: "Add review learning",
-    url: "https://example.test/pr/22",
-    state: "OPEN",
-    createdAt: "2026-03-11T14:00:00Z",
-    isDraft: false,
-    reviewDecision: "CHANGES_REQUESTED",
-    mergeStateStatus: "CLEAN",
-    headRefName: "codex/issue-58",
-    headRefOid: "deadbeef",
-  };
-
-  const status = formatDetailedStatus({
-    config,
-    activeRecord: createRecord({
-      pr_number: 22,
-      state: "addressing_review",
-      external_review_head_sha: "deadbeef",
-      external_review_misses_path: "/tmp/reviews/owner-repo/issue-58/external-review-misses-head-deadbeef.json",
-      external_review_matched_findings_count: 1,
-      external_review_near_match_findings_count: 1,
-      external_review_missed_findings_count: 2,
-    }),
-    latestRecord: null,
-    trackedIssueCount: 1,
-    pr,
-    checks: [],
-    reviewThreads: [],
-  });
-
-  assert.match(status, /external_review head=current reviewed_head_sha=deadbeef matched=1 near_match=1 missed=2/);
-  assert.match(status, /external_review_misses_path=owner-repo\/issue-58\/external-review-misses-head-deadbeef\.json/);
-});
-
-test("formatDetailedStatus surfaces not_requested Copilot review lifecycle", () => {
-  const config = createConfig({ copilotReviewWaitMinutes: 10 });
-  const pr: GitHubPullRequest = {
-    number: 22,
-    title: "Add review learning",
-    url: "https://example.test/pr/22",
-    state: "OPEN",
-    createdAt: "2026-03-11T14:00:00Z",
-    isDraft: false,
-    reviewDecision: null,
-    mergeStateStatus: "CLEAN",
-    headRefName: "codex/issue-58",
-    headRefOid: "deadbeef",
-  };
-
-  const status = formatDetailedStatus({
-    config,
-    activeRecord: createRecord({
-      pr_number: 22,
-      state: "pr_open",
-      review_wait_started_at: new Date().toISOString(),
-      review_wait_head_sha: "deadbeef",
-    }),
-    latestRecord: null,
-    trackedIssueCount: 1,
-    pr,
-    checks: [],
-    reviewThreads: [],
-  });
-
-  assert.match(status, /copilot_review state=not_requested/);
-});
-
-test("formatDetailedStatus surfaces unknown Copilot review lifecycle when hydration fails", () => {
-  const config = createConfig({ copilotReviewWaitMinutes: 10 });
-  const pr: GitHubPullRequest = {
-    number: 22,
-    title: "Add review learning",
-    url: "https://example.test/pr/22",
-    state: "OPEN",
-    createdAt: "2026-03-11T14:00:00Z",
-    isDraft: false,
-    reviewDecision: null,
-    mergeStateStatus: "CLEAN",
-    headRefName: "codex/issue-58",
-    headRefOid: "deadbeef",
-    copilotReviewState: null,
-    copilotReviewRequestedAt: null,
-    copilotReviewArrivedAt: null,
-  };
-
-  const status = formatDetailedStatus({
-    config,
-    activeRecord: createRecord({
-      pr_number: 22,
-      state: "pr_open",
-    }),
-    latestRecord: null,
-    trackedIssueCount: 1,
-    pr,
-    checks: [],
-    reviewThreads: [],
-  });
-
-  assert.match(status, /copilot_review state=unknown requested_at=none arrived_at=none/);
-});
-
-test("formatDetailedStatus surfaces Copilot review timeout outcome", () => {
-  const config = createConfig({ copilotReviewWaitMinutes: 10, copilotReviewTimeoutAction: "continue" });
-  const pr: GitHubPullRequest = {
-    number: 22,
-    title: "Add review learning",
-    url: "https://example.test/pr/22",
-    state: "OPEN",
-    createdAt: "2026-03-11T14:00:00Z",
-    isDraft: false,
-    reviewDecision: null,
-    mergeStateStatus: "CLEAN",
-    headRefName: "codex/issue-58",
-    headRefOid: "deadbeef",
-    copilotReviewState: "requested",
-    copilotReviewRequestedAt: "2026-03-11T14:05:00Z",
-    copilotReviewArrivedAt: null,
-  };
-
-  const status = formatDetailedStatus({
-    config,
-    activeRecord: createRecord({
-      pr_number: 22,
-      state: "ready_to_merge",
-      copilot_review_timed_out_at: "2026-03-11T14:15:00Z",
-      copilot_review_timeout_action: "continue",
-      copilot_review_timeout_reason:
-        "Requested Copilot review never arrived within 10 minute(s) for head deadbeef.",
-    }),
-    latestRecord: null,
-    trackedIssueCount: 1,
-    pr,
-    checks: [],
-    reviewThreads: [],
-  });
-
-  assert.match(status, /copilot_review state=requested requested_at=2026-03-11T14:05:00Z arrived_at=none timed_out_at=2026-03-11T14:15:00Z timeout_action=continue/);
-  assert.match(status, /timeout_reason=Requested Copilot review never arrived within 10 minute\(s\) for head deadbeef\./);
-});
-
-test("formatDetailedStatus surfaces configured bot review timeout outcome with generic wording", () => {
-  const config = createConfig({
-    reviewBotLogins: ["chatgpt-codex-connector"],
-  });
-  const pr: GitHubPullRequest = {
-    number: 44,
-    title: "Test PR",
-    url: "https://example.test/pr/44",
-    state: "OPEN",
-    createdAt: "2026-03-11T14:00:00Z",
-    isDraft: false,
-    reviewDecision: null,
-    mergeStateStatus: "CLEAN",
-    mergeable: "MERGEABLE",
-    headRefName: "codex/issue-38",
-    headRefOid: "deadbeef",
-    mergedAt: null,
-    copilotReviewState: "requested",
-    copilotReviewRequestedAt: "2026-03-11T14:05:00Z",
-    copilotReviewArrivedAt: null,
-  };
-  const status = formatDetailedStatus({
-    config,
-    activeRecord: createRecord({
-      pr_number: 44,
-      state: "blocked",
-      blocked_reason: "review_bot_timeout",
-      copilot_review_timed_out_at: "2026-03-11T14:15:00Z",
-      copilot_review_timeout_action: "continue",
-      copilot_review_timeout_reason:
-        "Requested configured review bot (chatgpt-codex-connector) review never arrived within 10 minute(s) for head deadbeef.",
-    }),
-    latestRecord: null,
-    trackedIssueCount: 1,
-    pr,
-    checks: [{ name: "build", state: "SUCCESS", bucket: "pass", workflow: "CI" }],
-    reviewThreads: [],
-  });
-
-  assert.match(
-    status,
-    /configured_bot_review state=requested reviewers=chatgpt-codex-connector requested_at=2026-03-11T14:05:00Z arrived_at=none timed_out_at=2026-03-11T14:15:00Z timeout_action=continue/,
-  );
-  assert.match(
-    status,
-    /timeout_reason=Requested configured review bot \(chatgpt-codex-connector\) review never arrived within 10 minute\(s\) for head deadbeef\./,
-  );
-});
-
-test("formatDetailedStatus explains softened nitpick-only configured-bot top-level reviews", () => {
-  const config = createConfig({
-    reviewBotLogins: ["coderabbitai[bot]"],
-  });
-  const pr: GitHubPullRequest = {
-    number: 44,
-    title: "Test PR",
-    url: "https://example.test/pr/44",
-    state: "OPEN",
-    createdAt: "2026-03-11T14:00:00Z",
-    isDraft: false,
-    reviewDecision: "CHANGES_REQUESTED",
-    mergeStateStatus: "CLEAN",
-    mergeable: "MERGEABLE",
-    headRefName: "codex/issue-38",
-    headRefOid: "deadbeef",
-    mergedAt: null,
-    copilotReviewState: "arrived",
-    copilotReviewRequestedAt: "2026-03-11T14:05:00Z",
-    copilotReviewArrivedAt: "2026-03-11T14:06:00Z",
-    configuredBotTopLevelReviewStrength: "nitpick_only",
-    configuredBotTopLevelReviewSubmittedAt: "2026-03-11T14:06:00Z",
-  };
-
-  const status = formatDetailedStatus({
-    config,
-    activeRecord: createRecord({
-      pr_number: 44,
-      state: "ready_to_merge",
-    }),
-    latestRecord: null,
-    trackedIssueCount: 1,
-    pr,
-    checks: [{ name: "build", state: "SUCCESS", bucket: "pass", workflow: "CI" }],
-    reviewThreads: [],
-  });
-
-  assert.match(
-    status,
-    /configured_bot_top_level_review strength=nitpick_only submitted_at=2026-03-11T14:06:00Z effect=softened/,
-  );
-});
-
-test("formatDetailedStatus surfaces active review-bot profile and missing external signal for Codex profile", () => {
-  const config = createConfig({
-    reviewBotLogins: ["chatgpt-codex-connector"],
-  });
-  const pr: GitHubPullRequest = {
-    number: 44,
-    title: "Test PR",
-    url: "https://example.test/pr/44",
-    state: "OPEN",
-    createdAt: "2026-03-11T14:00:00Z",
-    isDraft: false,
-    reviewDecision: "REVIEW_REQUIRED",
-    mergeStateStatus: "CLEAN",
-    mergeable: "MERGEABLE",
-    headRefName: "codex/issue-38",
-    headRefOid: "deadbeef",
-    mergedAt: null,
-    copilotReviewState: "not_requested",
-    copilotReviewRequestedAt: null,
-    copilotReviewArrivedAt: null,
-  };
-
-  const status = formatDetailedStatus({
-    config,
-    activeRecord: createRecord({
-      pr_number: 44,
-      state: "pr_open",
-      blocked_reason: null,
-      external_review_head_sha: null,
-      external_review_matched_findings_count: 0,
-      external_review_near_match_findings_count: 0,
-      external_review_missed_findings_count: 0,
-      copilot_review_timed_out_at: null,
-      copilot_review_timeout_action: null,
-      copilot_review_timeout_reason: null,
-    }),
-    latestRecord: null,
-    trackedIssueCount: 1,
-    pr,
-    checks: [{ name: "build", state: "SUCCESS", bucket: "pass", workflow: "CI" }],
-    reviewThreads: [],
-  });
-
-  assert.match(
-    status,
-    /review_bot_profile profile=codex provider=chatgpt-codex-connector reviewers=chatgpt-codex-connector signal_source=review_threads/,
-  );
-  assert.match(
-    status,
-    /review_bot_diagnostics status=missing_provider_signal observed_review=none expected_reviewers=chatgpt-codex-connector next_check=provider_setup_or_delivery/,
-  );
-});
-
-test("formatDetailedStatus detects the CodeRabbit profile for canonical and reversed review bot login orderings", () => {
-  const pr: GitHubPullRequest = {
-    number: 44,
-    title: "Test PR",
-    url: "https://example.test/pr/44",
-    state: "OPEN",
-    createdAt: "2026-03-11T14:00:00Z",
-    isDraft: false,
-    reviewDecision: "REVIEW_REQUIRED",
-    mergeStateStatus: "CLEAN",
-    mergeable: "MERGEABLE",
-    headRefName: "codex/issue-38",
-    headRefOid: "deadbeef",
-    mergedAt: null,
-    copilotReviewState: "not_requested",
-    copilotReviewRequestedAt: null,
-    copilotReviewArrivedAt: null,
-  };
-
-  for (const reviewBotLogins of [
-    ["coderabbitai", "coderabbitai[bot]"],
-    ["coderabbitai[bot]", "coderabbitai"],
-  ]) {
-    const config = createConfig({ reviewBotLogins });
-    const status = formatDetailedStatus({
-      config,
-      activeRecord: createRecord({
-        pr_number: 44,
-        state: "pr_open",
-        blocked_reason: null,
-        external_review_head_sha: null,
-        external_review_matched_findings_count: 0,
-        external_review_near_match_findings_count: 0,
-        external_review_missed_findings_count: 0,
-        copilot_review_timed_out_at: null,
-        copilot_review_timeout_action: null,
-        copilot_review_timeout_reason: null,
-      }),
-      latestRecord: null,
-      trackedIssueCount: 1,
-      pr,
-      checks: [{ name: "build", state: "SUCCESS", bucket: "pass", workflow: "CI" }],
-      reviewThreads: [],
-    });
-
-    assert.match(
-      status,
-      /review_bot_profile profile=coderabbit provider=coderabbitai .* signal_source=review_threads/,
-    );
-  }
-});
-
-test("formatDetailedStatus preserves Copilot-specific timeout wording for Copilot-only repos", () => {
-  const config = createConfig({
-    reviewBotLogins: ["copilot-pull-request-reviewer"],
-  });
-  const pr: GitHubPullRequest = {
-    number: 22,
-    title: "Add review learning",
-    url: "https://example.test/pr/22",
-    state: "OPEN",
-    createdAt: "2026-03-11T14:00:00Z",
-    isDraft: false,
-    reviewDecision: null,
-    mergeStateStatus: "CLEAN",
-    headRefName: "codex/issue-58",
-    headRefOid: "deadbeef",
-    copilotReviewState: "requested",
-    copilotReviewRequestedAt: "2026-03-11T14:05:00Z",
-    copilotReviewArrivedAt: null,
-  };
-
-  const status = formatDetailedStatus({
-    config,
-    activeRecord: createRecord({
-      pr_number: 22,
-      state: "ready_to_merge",
-      copilot_review_timed_out_at: "2026-03-11T14:15:00Z",
-      copilot_review_timeout_action: "continue",
-      copilot_review_timeout_reason:
-        "Requested Copilot review never arrived within 10 minute(s) for head deadbeef.",
-    }),
-    latestRecord: null,
-    trackedIssueCount: 1,
-    pr,
-    checks: [],
-    reviewThreads: [],
-  });
-
-  assert.match(status, /copilot_review state=requested requested_at=2026-03-11T14:05:00Z arrived_at=none timed_out_at=2026-03-11T14:15:00Z timeout_action=continue/);
-  assert.match(status, /timeout_reason=Requested Copilot review never arrived within 10 minute\(s\) for head deadbeef\./);
-});
-
-test("formatDetailedStatus keeps generic configured bot timeout wording for mixed bot repos", () => {
-  const config = createConfig({
-    reviewBotLogins: ["copilot-pull-request-reviewer", "chatgpt-codex-connector"],
-  });
-  const pr: GitHubPullRequest = {
-    number: 44,
-    title: "Test PR",
-    url: "https://example.test/pr/44",
-    state: "OPEN",
-    createdAt: "2026-03-11T14:00:00Z",
-    isDraft: false,
-    reviewDecision: null,
-    mergeStateStatus: "CLEAN",
-    mergeable: "MERGEABLE",
-    headRefName: "codex/issue-38",
-    headRefOid: "deadbeef",
-    mergedAt: null,
-    copilotReviewState: "requested",
-    copilotReviewRequestedAt: "2026-03-11T14:05:00Z",
-    copilotReviewArrivedAt: null,
-  };
-  const status = formatDetailedStatus({
-    config,
-    activeRecord: createRecord({
-      pr_number: 44,
-      state: "blocked",
-      blocked_reason: "review_bot_timeout",
-      copilot_review_timed_out_at: "2026-03-11T14:15:00Z",
-      copilot_review_timeout_action: "continue",
-      copilot_review_timeout_reason:
-        "Requested configured review bots (copilot-pull-request-reviewer, chatgpt-codex-connector) review never arrived within 10 minute(s) for head deadbeef.",
-    }),
-    latestRecord: null,
-    trackedIssueCount: 1,
-    pr,
-    checks: [{ name: "build", state: "SUCCESS", bucket: "pass", workflow: "CI" }],
-    reviewThreads: [],
-  });
-
-  assert.match(
-    status,
-    /configured_bot_review state=requested reviewers=copilot-pull-request-reviewer,chatgpt-codex-connector requested_at=2026-03-11T14:05:00Z arrived_at=none timed_out_at=2026-03-11T14:15:00Z timeout_action=continue/,
-  );
-  assert.match(
-    status,
-    /timeout_reason=Requested configured review bots \(copilot-pull-request-reviewer, chatgpt-codex-connector\) review never arrived within 10 minute\(s\) for head deadbeef\./,
-  );
-});
-
-test("formatDetailedStatus surfaces the latest recovery reason separately from the active issue", () => {
-  const config = createConfig();
-  const activeRecord = createRecord({
-    issue_number: 92,
-    state: "implementing",
-    branch: "codex/issue-92",
-    workspace: "/tmp/workspaces/issue-92",
-    blocked_reason: null,
-    last_error: null,
-    last_failure_kind: null,
-    last_failure_context: null,
-    last_failure_signature: null,
-    codex_session_id: null,
-  });
-  const latestRecoveryRecord = createRecord({
-    issue_number: 91,
-    state: "done",
-    branch: "codex/issue-91",
-    workspace: "/tmp/workspaces/issue-91",
-    blocked_reason: null,
-    last_error: null,
-    last_failure_kind: null,
-    last_failure_context: null,
-    last_failure_signature: null,
-    codex_session_id: null,
-    updated_at: "2026-03-13T00:20:00Z",
-    last_codex_summary: null,
-    last_recovery_reason: "merged_pr_convergence: tracked PR #191 merged; marked issue #91 done",
-    last_recovery_at: "2026-03-13T00:20:00Z",
-  });
-
-  const status = formatDetailedStatus({
-    config,
-    activeRecord,
-    latestRecord: latestRecoveryRecord,
-    latestRecoveryRecord,
-    trackedIssueCount: 2,
-    pr: null,
-    checks: [],
-    reviewThreads: [],
-  });
-
-  assert.match(
-    status,
-    /latest_recovery issue=#91 at=2026-03-13T00:20:00Z reason=merged_pr_convergence: tracked PR #191 merged; marked issue #91 done/,
-  );
-});
-
-test("formatDetailedStatus reports idle status with the latest record and latest recovery", () => {
-  const config = createConfig();
-  const latestRecord = createRecord({
-    issue_number: 92,
-    state: "done",
-    branch: "codex/issue-92",
-    updated_at: "2026-03-13T01:20:00Z",
-  });
-  const latestRecoveryRecord = createRecord({
-    issue_number: 91,
-    state: "done",
-    branch: "codex/issue-91",
-    workspace: "/tmp/workspaces/issue-91",
-    updated_at: "2026-03-13T00:20:00Z",
-    last_codex_summary: null,
-    last_recovery_reason: "merged_pr_convergence: tracked PR #191 merged; marked issue #91 done",
-    last_recovery_at: "2026-03-13T00:20:00Z",
-  });
-
-  const status = formatDetailedStatus({
-    config,
-    activeRecord: null,
-    latestRecord,
-    latestRecoveryRecord,
-    trackedIssueCount: 2,
-    pr: null,
-    checks: [],
-    reviewThreads: [],
-  });
-
-  assert.match(status, /^No active issue\./);
-  assert.match(status, /tracked_issues=2/);
-  assert.match(status, /latest_record=#92 state=done updated_at=2026-03-13T01:20:00Z/);
-  assert.match(
-    status,
-    /latest_recovery issue=#91 at=2026-03-13T00:20:00Z reason=merged_pr_convergence: tracked PR #191 merged; marked issue #91 done/,
-  );
-});
-
-test("configuredBotReviewThreads normalizes configured bot logins before classifying threads", () => {
-  const config = createConfig({
-    reviewBotLogins: [" Copilot-Pull-Request-Reviewer "],
-  });
-  const thread = createReviewThread({
-    comments: {
-      nodes: [
-        {
-          id: "comment-1",
-          body: "Please address this.",
-          createdAt: "2026-03-11T00:00:00Z",
-          url: "https://example.test/pr/44#discussion_r1",
-          author: {
-            login: "copilot-pull-request-reviewer",
-            typeName: "Bot",
-          },
-        },
-      ],
-    },
-  });
-
-  assert.equal(configuredBotReviewThreads(config, [thread]).length, 1);
-  assert.equal(manualReviewThreads(config, [thread]).length, 0);
-});
-
-test("formatDetailedStatus marks stale local review as non-gating", () => {
-  const config = createConfig({ localReviewPolicy: "block_merge" });
-  const record = createRecord({
-    local_review_head_sha: "oldhead",
-    local_review_max_severity: "medium",
-    local_review_findings_count: 2,
-    local_review_recommendation: "changes_requested",
-    local_review_run_at: "2026-03-11T14:05:00Z",
-  });
-  const pr: GitHubPullRequest = {
-    number: 42,
-    title: "Test PR",
-    url: "https://example.test/pr/42",
-    state: "OPEN",
-    createdAt: "2026-03-11T14:00:00Z",
-    isDraft: false,
-    reviewDecision: null,
-    mergeStateStatus: "CLEAN",
-    mergeable: "MERGEABLE",
-    headRefName: "codex/issue-42",
-    headRefOid: "newhead",
-    mergedAt: null,
-  };
-
-  const status = formatDetailedStatus({
-    config,
-    activeRecord: record,
-    latestRecord: record,
-    trackedIssueCount: 1,
-    pr,
-    checks: [],
-    reviewThreads: [],
-  });
-
-  assert.match(
-    status,
-    /local_review gating=no policy=block_merge findings=2 root_causes=0 max_severity=medium verified_findings=0 verified_max_severity=none head=stale reviewed_head_sha=oldhead pr_head_sha=newhead ran_at=2026-03-11T14:05:00Z needs_review_run=yes drift=oldhead->newhead/,
-  );
-});
-
-test("formatDetailedStatus reports unknown local review head status without a PR", () => {
-  const config = createConfig({ localReviewPolicy: "block_merge" });
-  const record = createRecord({
-    local_review_head_sha: "oldhead",
-    local_review_max_severity: "medium",
-    local_review_findings_count: 2,
-    local_review_recommendation: "changes_requested",
-    local_review_run_at: "2026-03-11T14:05:00Z",
-  });
-
-  const status = formatDetailedStatus({
-    config,
-    activeRecord: record,
-    latestRecord: record,
-    trackedIssueCount: 1,
-    pr: null,
-    checks: [],
-    reviewThreads: [],
-  });
-
-  assert.match(
-    status,
-    /local_review gating=no policy=block_merge findings=2 root_causes=0 max_severity=medium verified_findings=0 verified_max_severity=none head=unknown reviewed_head_sha=oldhead pr_head_sha=unknown ran_at=2026-03-11T14:05:00Z/,
-  );
-});
-
-test("formatDetailedStatus reports none local review head status with current PR head", () => {
-  const config = createConfig({ localReviewPolicy: "block_merge" });
-  const record = createRecord({
-    local_review_head_sha: null,
-    local_review_run_at: null,
-  });
-  const pr: GitHubPullRequest = {
-    number: 42,
-    title: "Test PR",
-    url: "https://example.test/pr/42",
-    state: "OPEN",
-    createdAt: "2026-03-11T14:00:00Z",
-    isDraft: false,
-    reviewDecision: null,
-    mergeStateStatus: "CLEAN",
-    mergeable: "MERGEABLE",
-    headRefName: "codex/issue-42",
-    headRefOid: "newhead",
-    mergedAt: null,
-  };
-
-  const status = formatDetailedStatus({
-    config,
-    activeRecord: record,
-    latestRecord: record,
-    trackedIssueCount: 1,
-    pr,
-    checks: [],
-    reviewThreads: [],
-  });
-
-  assert.match(
-    status,
-    /local_review gating=no policy=block_merge findings=0 root_causes=0 max_severity=none verified_findings=0 verified_max_severity=none head=none reviewed_head_sha=none pr_head_sha=newhead ran_at=none/,
-  );
-  assert.doesNotMatch(status, /blocker_summary=/);
-});
-
-test("localReviewHighSeverityNeedsRetry only escalates verifier-confirmed high findings", () => {
-  const config = createConfig({ localReviewPolicy: "block_ready", localReviewHighSeverityAction: "retry" });
-  const pr: GitHubPullRequest = {
-    number: 42,
-    title: "Test PR",
-    url: "https://example.test/pr/42",
-    state: "OPEN",
-    createdAt: "2026-03-11T14:00:00Z",
-    isDraft: true,
-    reviewDecision: null,
-    mergeStateStatus: "CLEAN",
-    mergeable: "MERGEABLE",
-    headRefName: "codex/issue-42",
-    headRefOid: "deadbeef",
-    mergedAt: null,
-  };
-
-  assert.equal(
-    localReviewHighSeverityNeedsRetry(
-      config,
-      {
-        local_review_head_sha: "deadbeef",
-        local_review_verified_max_severity: "none",
-      },
-      pr,
-    ),
-    false,
-  );
-
-  assert.equal(
-    localReviewHighSeverityNeedsRetry(
-      config,
-      {
-        local_review_head_sha: "deadbeef",
-        local_review_verified_max_severity: "high",
-      },
-      pr,
-    ),
-    true,
-  );
 });
